Awesome map
I added quit a bit of logic to handle move, search, click, ...

By changing collapseAfterResult & expanded options, it's even possible to completely get rid of #sortie_RDV
But I prefer to keep it for now, in case a user wants to add "location explanations". We need to store both the comment and the marker point (position/address) in fact.