Event location improvements
- Never use event location for map marker
- Always show a marker of the rendez-vous point (a map without marker is useless)
Thus, default rendez-vous point should be displayed when nothing was changed
In theory, we should save the gps & gps_location in the database for each new event. Because the default location might change with time.
However, this would add uneeded data transfer when getting the list of events. And we don't really care the get the exact rendez-vous location for (very old) events.